rename tests to match other integration tests
